Separate logic for current Login component
