+++ .env.example
@@ -14,13 +14,13 @@ REACT_APP_URL_SHORTENER_DOMAIN=
 # for the correct auth details
 REACT_APP_OAUTH2_PROVIDER1=test_admin
 REACT_APP_OAUTH2_PROVIDER1_NAME=Test login as admin
-REACT_APP_OAUTH2_PROVIDER1_ENDPOINT=http://localhost:8000/login?provider=test&code=admin_code
+REACT_APP_OAUTH2_PROVIDER1_ENDPOINT=http://localhost:8000/login/callback?provider=test&code=admin_code
 REACT_APP_OAUTH2_PROVIDER2=test_staff
 REACT_APP_OAUTH2_PROVIDER2_NAME=Test login as staff
-REACT_APP_OAUTH2_PROVIDER2_ENDPOINT=http://localhost:8000/login?provider=test&code=staff_code
+REACT_APP_OAUTH2_PROVIDER2_ENDPOINT=http://localhost:8000/login/callback?provider=test&code=staff_code
 REACT_APP_OAUTH2_PROVIDER3=test_student
 REACT_APP_OAUTH2_PROVIDER3_NAME=Test login as student
-REACT_APP_OAUTH2_PROVIDER3_ENDPOINT=http://localhost:8000/login?provider=test&code=student_code
+REACT_APP_OAUTH2_PROVIDER3_ENDPOINT=http://localhost:8000/login/callback?provider=test&code=student_code
 
 ## LumiNUS example
 ## the provider ID, must be URL-friendly (must match the backend configuration)
